Number is now a float type

+++ packages/@tinacms/graphql/src/ast-builder/index.ts
@@ -534,7 +534,7 @@ export const astBuilder = {
     Node: 'Node',
     PageInfo: 'PageInfo',
     Connection: 'Connection',
-    Number: 'Int',
+    Number: 'Float',
     Document: 'Document',
   },
